Move components in respective subfolders
