+++ src/components/Canvas/DraggingComponent.js
@@ -11,9 +11,9 @@ import React, { Component } from 'react';
 //import {
   //throttle,
 //} from 'lodash-es';
-import Mouse from '../utils/mouse';
-import Canvas from '../drawing/canvas';
-import WebGL from '../drawing/webgl';
+import Mouse from '../../utils/mouse';
+import Canvas from '../../drawing/canvas';
+import WebGL from '../../drawing/webgl';
 
 import ModBar from './ModBar';
