+++ src/components/Canvas/Labels.js
