+++ src/components/Canvas/SequenceViewer.js
