+++ src/components/SequenceViewer/Residue.js
