+++ src/components/SequenceViewer/Sequence.js
