+++ src/components/index.js
@@ -1,0 +1,20 @@
+/**
+* Copyright 2018, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+import Labels from './yBars/Labels';
+import PositionBar from './xBars/PositionBar';
+import OverviewBar from './xBars/OverviewBar';
+import SequenceViewer from './SequenceViewer/SequenceViewer';
+import SequenceOverview from './Canvas/SequenceOverview';
+export {
+  Labels,
+  PositionBar,
+  OverviewBar,
+  SequenceViewer,
+  SequenceOverview,
+};
